+++ src/Snackbar/index.ts
@@ -1,5 +1,5 @@
 /* ======================================================================== *
- * Copyright 2024 HCL America Inc.                                          *
+ * Copyright 2024, 2026 HCL America Inc.                                    *
  * Licensed under the Apache License, Version 2.0 (the "License");          *
  * you may not use this file except in compliance with the License.         *
  * You may obtain a copy of the License at                                  *
@@ -16,3 +16,5 @@ import Snackbar from './Snackbar';
 
 export default Snackbar;
 export * from './Snackbar';
+export { default as GroupedSnackbar } from './GroupedSnackbar';
+export * from './GroupedSnackbar';